added test setup
